requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django>=4.0.3
django-email-users==1.1.0
django-environ==0.8.1
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
drf-yasg==1.20.0
environ==1.0
factory-boy==3.2.1
Faker==13.3.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
